pyproject.toml
[tool.poetry]
name = "django-challenge"
version = "0.1.0"
description = ""
authors = ["EmannuelDM"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.2"
markdown = "^3.6"
pytest = "^8.2.2"
django-filter = "^24.2"
django-extensions = "^3.2.3"
drf-yasg = "^1.21.7"
whitenoise = "^6.7.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
pytest-django = "^4.8.0"
pre-commit = "^3.8.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
